requirements.txt
gdown>=4.7.1
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.2.2
pillow==11.1.0
tqdm==4.67.1
torch==2.6.0+cu124
torchvision==0.21.0+cu124
transformers==4.51.3
datasets==3.6.0
evaluate==0.4.3
peft==0.14.0
sentence-transformers==3.4.1
